refactor: Extract row actions and confirmation dialog from profile-tables
